TransformationCleaningAgent: add option to submit removeFile requests instead of direct cleanup

+++ TransformationSystem/Agent/TransformationCleaningAgent.py
@@ -14,7 +14,10 @@ __RCSID__ = "$Id$"
 import re
 import ast
 import os.path
+from hashlib import md5
 from datetime import datetime, timedelta
+import time
+
 # # from DIRAC
 from DIRAC import S_OK, S_ERROR
 from DIRAC.Core.Base.AgentModule import AgentModule
@@ -30,6 +33,11 @@ from DIRAC.Core.Utilities.ReturnValues import returnSingleResult
 from DIRAC.Resources.Catalog.FileCatalog import FileCatalog
 from DIRAC.ConfigurationSystem.Client.ConfigurationData import gConfigurationData
 from DIRAC.RequestManagementSystem.Client.ReqClient import ReqClient
+from DIRAC.RequestManagementSystem.Client.Request import Request
+from DIRAC.RequestManagementSystem.Client.Operation import Operation
+from DIRAC.RequestManagementSystem.Client.File import File
+from DIRAC.RequestManagementSystem.private.RequestValidator import RequestValidator
+
 
 # # agent's name
 AGENT_NAME = 'Transformation/TransformationCleaningAgent'
@@ -115,6 +123,7 @@ class TransformationCleaningAgent(AgentModule):
     # # transformation log SEs
     self.logSE = Operations().getValue('/LogStorage/LogSE', self.logSE)
     self.log.info("Will remove logs found on storage element: %s" % self.logSE)
+    self.cleanWithRMS = self.am_getOption('CleanWithRMS', True)
 
     # # transformation client
     self.transClient = TransformationClient()
@@ -348,9 +357,12 @@ class TransformationCleaningAgent(AgentModule):
     self.log.info("Attempting to remove %d possible remnants from the catalog and storage" % len(filesFound))
 
     # Executing with shifter proxy
-    gConfigurationData.setOptionInCFG('/DIRAC/Security/UseServerCertificate', 'false')
-    res = DataManager().removeFile(filesFound, force=True)
-    gConfigurationData.setOptionInCFG('/DIRAC/Security/UseServerCertificate', 'true')
+    if self.cleanWithRMS:
+      res = self.__submitRemovalRequests(0, filesFound)
+    else:
+      gConfigurationData.setOptionInCFG('/DIRAC/Security/UseServerCertificate', 'false')
+      res = DataManager().removeFile(filesFound, force=True)
+      gConfigurationData.setOptionInCFG('/DIRAC/Security/UseServerCertificate', 'true')
 
     if not res['OK']:
       return res
@@ -521,9 +533,12 @@ class TransformationCleaningAgent(AgentModule):
       return S_OK()
 
     # Executing with shifter proxy
-    gConfigurationData.setOptionInCFG('/DIRAC/Security/UseServerCertificate', 'false')
-    res = DataManager().removeFile(fileToRemove, force=True)
-    gConfigurationData.setOptionInCFG('/DIRAC/Security/UseServerCertificate', 'true')
+    if self.cleanWithRMS:
+      res = self.__submitRemovalRequests(transID, fileToRemove)
+    else:
+      gConfigurationData.setOptionInCFG('/DIRAC/Security/UseServerCertificate', 'false')
+      res = DataManager().removeFile(fileToRemove, force=True)
+      gConfigurationData.setOptionInCFG('/DIRAC/Security/UseServerCertificate', 'true')
 
     if not res['OK']:
       return res
@@ -654,4 +669,46 @@ class TransformationCleaningAgent(AgentModule):
       self.log.info("Failed to remove %s requests" % failed)
       return S_ERROR("Failed to remove all the request from RequestDB")
     self.log.info("Successfully removed all the associated failover requests")
+    return S_OK()
+
+  def __submitRemovalRequests(self, tID, lfns):
+    """Create removal requests for all files."""
+    for index, lfnList in enumerate(breakListIntoChunks(lfns, 300)):
+      oRequest = Request()
+      requestName = "TCA_%s_%s_%s" % (tID, index, md5(repr(time.time())).hexdigest()[:5])
+      oRequest.RequestName = requestName
+      oOperation = Operation()
+      oOperation.Type = 'RemoveFile'
+      oOperation.TargetSE = 'All'
+      res = self.metadataClient.getFileMetadata(lfnList)
+      if not res['OK']:
+        self.log.error('Cannot get file metadata', res['Message'])
+        return res
+      if res['Value']['Failed']:
+        self.log.warning('Could not get the file metadata of the following, so skipping them:',
+                         res['Value']['Failed'])
+      lfnMetadata = res['Value']['Successful']
+
+      for lfn in lfnMetadata:
+        rarFile = File()
+        rarFile.LFN = lfn
+        rarFile.Size = lfnMetadata[lfn]['Size']
+        rarFile.Checksum = lfnMetadata[lfn]['Checksum']
+        rarFile.GUID = lfnMetadata[lfn]['GUID']
+        rarFile.ChecksumType = 'ADLER32'
+        oOperation.addFile(rarFile)
+
+      oRequest.addOperation(oOperation)
+      isValid = RequestValidator().validate(oRequest)
+      if not isValid['OK']:
+        self.log.error('Request is not valid:', isValid['Message'])
+        return isValid
+      result = self.reqClient.putRequest(oRequest)
+      if result['OK']:
+        self.log.info('RemoveFiles request %d submitted for %d LFNs' %
+                      (result['Value'], len(lfnMetadata)))
+      else:
+        self.log.error('Failed to submit Request:', result['Message'])
+        return result
+
     return S_OK()
